fix some edge cases

+++ users/common/gnome/gnome-wallpaper-fetch.sh
@@ -1,10 +1,18 @@
 mkdir -p ~/.wallpaper
 cd ~/.wallpaper
 oldFile=$(ls)
+if [ "$oldFile" = "" ]
+then
+    oldFile="0.jpg"
+fi
 filename=$(basename "$oldFile")
 filename=${filename%.*}
 newFile="$(($filename + 1)).jpg"
-
+if [ "$newFile" = "" ]
+then
+    echo "Error"
+    exit 1
+fi
 echo "Fetching index"
 baseURL="https://raw.githubusercontent.com/clayts/Wallpapers/master"
 indexPath="index.txt"
@@ -27,6 +35,9 @@ temp=$(mktemp) &&
 
 echo "Executing: gsettings set org.gnome.desktop.background picture-uri '.wallpaper/$newFile'"
 gsettings set org.gnome.desktop.background picture-uri ".wallpaper/$newFile"
-sleep 2s
-echo "Removing ~/.wallpaper/$oldFile"
-rm "$oldFile"
+if [ ! "$oldFile" = "0.jpg" ]
+then
+    sleep 2s
+    echo "Removing ~/.wallpaper/$oldFile"
+    rm "$oldFile"
+fi
